django-backend/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.6.0
django-model-utils==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
whitenoise==3.3.1
django-heroku==0.3.1